fix the STM32L475 recv method (long msg error)

+++ B-L475E-IOT01/acnsdkc/arrow/storage.c
@@ -15,7 +15,7 @@
 flash_mem_t mem __attribute__((section("UNINIT_FIXED_LOC")));
 
 int restore_gateway_info(arrow_gateway_t *gateway) {
-  if ( mem.magic != FLASH_MAGIC_NUMBER ) {
+  if ( mem.magic != (int) FLASH_MAGIC_NUMBER ) {
     FLASH_unlock_erase((uint32_t)&mem, sizeof(mem));
     return -1;
   }
@@ -40,7 +40,7 @@ void save_gateway_info(const arrow_gateway_t *gateway) {
 }
 
 int restore_device_info(arrow_device_t *device) {
-  if ( mem.magic != FLASH_MAGIC_NUMBER ) {
+  if ( mem.magic != (int) FLASH_MAGIC_NUMBER ) {
     FLASH_unlock_erase((uint32_t)&mem, sizeof(mem));
     return -1;
   }
